# Optimize boolean value parsing