--- a/bluejay-parser/src/ast/value.rs
+++ b/bluejay-parser/src/ast/value.rs
@@ -1,5 +1,5 @@
 use crate::ast::{FromTokens, ParseError, Tokens, TryFromTokens, Variable};
-use crate::lexical_token::{FloatValue, IntValue, Name, PunctuatorType, StringValue};
+use crate::lexical_token::{BooleanValue, FloatValue, IntValue, Name, PunctuatorType, StringValue};
 use crate::{HasSpan, Span};
 use bluejay_core::{
     AsIter, ListValue as CoreListValue, ObjectValue as CoreObjectValue, Value as CoreValue,
@@ -41,92 +41,85 @@ impl<'a, const CONST: bool> CoreValue<CONST> for Value<'a, CONST> {
 
 impl<'a, const CONST: bool> FromTokens<'a> for Value<'a, CONST> {
     fn from_tokens(tokens: &mut impl Tokens<'a>) -> Result<Self, ParseError> {
-        None.or_else(|| {
-            if CONST {
-                None
-            } else {
-                Variable::try_from_tokens(tokens).map(|res| res.map(Self::Variable))
-            }
-        })
-        .or_else(|| {
-            tokens
-                .next_if_int_value()
-                .map(|i| Ok(Self::Integer(i)))
-                .or_else(|| tokens.next_if_float_value().map(|f| Ok(Self::Float(f))))
-                .or_else(|| tokens.next_if_string_value().map(|s| Ok(Self::String(s))))
-                .or_else(|| {
-                    tokens.next_if_name().map(|name| {
-                        Ok(match name.as_str() {
-                            "true" => Self::Boolean(BooleanValue {
-                                value: true,
-                                span: name.into(),
-                            }),
-                            "false" => Self::Boolean(BooleanValue {
-                                value: false,
-                                span: name.into(),
-                            }),
-                            "null" => Self::Null(name),
-                            _ => Self::Enum(name),
+        let result = if CONST {
+            None
+        } else {
+            Variable::try_from_tokens(tokens).map(|res| res.map(Self::Variable))
+        };
+
+        result
+            .or_else(|| {
+                tokens
+                    .next_if_int_value()
+                    .map(|i| Ok(Self::Integer(i)))
+                    .or_else(|| tokens.next_if_float_value().map(|f| Ok(Self::Float(f))))
+                    .or_else(|| tokens.next_if_string_value().map(|s| Ok(Self::String(s))))
+                    .or_else(|| tokens.next_if_boolean_value().map(|s| Ok(Self::Boolean(s))))
+                    .or_else(|| {
+                        tokens.next_if_name().map(|name| {
+                            Ok(match name.as_str() {
+                                "null" => Self::Null(name),
+                                _ => Self::Enum(name),
+                            })
                         })
                     })
-                })
-                .or_else(|| {
-                    tokens
-                        .next_if_punctuator(PunctuatorType::OpenSquareBracket)
-                        .map(|open_span| {
-                            let mut list: Vec<Self> = Vec::new();
-                            let close_span = loop {
-                                if let Some(close_span) =
-                                    tokens.next_if_punctuator(PunctuatorType::CloseSquareBracket)
-                                {
-                                    break close_span;
-                                }
-                                list.push(Self::from_tokens(tokens)?);
-                            };
-                            let span = open_span.merge(&close_span);
-                            Ok(Self::List(ListValue {
-                                elements: list,
-                                span,
-                            }))
-                        })
-                })
-                .or_else(|| {
-                    tokens
-                        .next_if_punctuator(PunctuatorType::OpenBrace)
-                        .map(|open_span| {
-                            let mut object: Vec<_> = Vec::new();
-                            let close_span = loop {
-                                if let Some(close_span) =
-                                    tokens.next_if_punctuator(PunctuatorType::CloseBrace)
-                                {
-                                    break close_span;
-                                }
-                                let name = tokens.expect_name()?;
-                                tokens.expect_punctuator(PunctuatorType::Colon)?;
-                                let value = Self::from_tokens(tokens)?;
-                                object.push((name, value));
-                            };
-                            let span = open_span.merge(&close_span);
-                            Ok(Self::Object(ObjectValue {
-                                fields: object,
-                                span,
-                            }))
-                        })
-                })
-        })
-        .unwrap_or_else(|| {
-            Err(tokens
-                .next()
-                .map(|token| ParseError::UnexpectedToken { span: token.into() })
-                .unwrap_or_else(|| tokens.unexpected_eof()))
-        })
+                    .or_else(|| {
+                        tokens
+                            .next_if_punctuator(PunctuatorType::OpenSquareBracket)
+                            .map(|open_span| {
+                                let mut list: Vec<Self> = Vec::new();
+                                let close_span = loop {
+                                    if let Some(close_span) = tokens
+                                        .next_if_punctuator(PunctuatorType::CloseSquareBracket)
+                                    {
+                                        break close_span;
+                                    }
+                                    list.push(Self::from_tokens(tokens)?);
+                                };
+                                let span = open_span.merge(&close_span);
+                                Ok(Self::List(ListValue {
+                                    elements: list,
+                                    span,
+                                }))
+                            })
+                    })
+                    .or_else(|| {
+                        tokens
+                            .next_if_punctuator(PunctuatorType::OpenBrace)
+                            .map(|open_span| {
+                                let mut object: Vec<_> = Vec::new();
+                                let close_span = loop {
+                                    if let Some(close_span) =
+                                        tokens.next_if_punctuator(PunctuatorType::CloseBrace)
+                                    {
+                                        break close_span;
+                                    }
+                                    let name = tokens.expect_name()?;
+                                    tokens.expect_punctuator(PunctuatorType::Colon)?;
+                                    let value = Self::from_tokens(tokens)?;
+                                    object.push((name, value));
+                                };
+                                let span = open_span.merge(&close_span);
+                                Ok(Self::Object(ObjectValue {
+                                    fields: object,
+                                    span,
+                                }))
+                            })
+                    })
+            })
+            .unwrap_or_else(|| {
+                Err(tokens
+                    .next()
+                    .map(|token| ParseError::UnexpectedToken { span: token.into() })
+                    .unwrap_or_else(|| tokens.unexpected_eof()))
+            })
     }
 }
 
 impl<'a, const CONST: bool> HasSpan for Value<'a, CONST> {
     fn span(&self) -> &Span {
         match self {
-            Self::Boolean(b) => &b.span,
+            Self::Boolean(b) => b.span(),
             Self::Enum(e) => e.span(),
             Self::Float(f) => f.span(),
             Self::Integer(i) => i.span(),
@@ -141,18 +134,6 @@ impl<'a, const CONST: bool> HasSpan for Value<'a, CONST> {
 
 pub type ConstValue<'a> = Value<'a, true>;
 pub type VariableValue<'a> = Value<'a, false>;
-
-#[derive(Debug)]
-pub struct BooleanValue {
-    value: bool,
-    span: Span,
-}
-
-impl BooleanValue {
-    fn value(&self) -> bool {
-        self.value
-    }
-}
 
 #[derive(Debug)]
 pub struct ListValue<'a, const CONST: bool> {