--- a/bluejay-parser/src/lexical_token.rs
+++ b/bluejay-parser/src/lexical_token.rs
@@ -1,11 +1,13 @@
 use crate::{HasSpan, Span};
 use enum_as_inner::EnumAsInner;
 
+mod bool_value;
 mod float_value;
 mod int_value;
 mod name;
 mod punctuator;
 mod string_value;
+pub use bool_value::BooleanValue;
 pub use float_value::FloatValue;
 pub use int_value::IntValue;
 pub use name::Name;
@@ -19,6 +21,7 @@ pub enum LexicalToken<'a> {
     IntValue(IntValue),
     FloatValue(FloatValue),
     StringValue(StringValue<'a>),
+    BooleanValue(BooleanValue),
 }
 
 impl<'a> HasSpan for LexicalToken<'a> {
@@ -29,6 +32,7 @@ impl<'a> HasSpan for LexicalToken<'a> {
             Self::StringValue(s) => s.span(),
             Self::Name(n) => n.span(),
             Self::Punctuator(p) => p.span(),
+            Self::BooleanValue(p) => p.span(),
         }
     }
 }
@@ -41,6 +45,7 @@ impl<'a> From<LexicalToken<'a>> for Span {
             LexicalToken::StringValue(s) => s.into(),
             LexicalToken::Name(n) => n.into(),
             LexicalToken::Punctuator(p) => p.into(),
+            LexicalToken::BooleanValue(p) => p.into(),
         }
     }
 }