Small tweak that makes us go from

```
time:   [11.633 µs 11.662 µs 11.696 µs]
```

to

```
time:   [11.223 µs 11.256 µs 11.293 µs]

```