@JoviDeCroock 2024-06-26
@adampetro that is a valid concern, hmm, might have to look for a different way to reduce the wrapping in our value conversion 😅